Refactor: board 패키지에서 다른 패키지로 파일 이동
board 패키지는 삭제 예정이므로 사용 중인 파일을 다른 패키지로 이동시킴
