+++ src/main/java/com/filmdoms/community/article/data/constant/PostStatus.java
@@ -1,0 +1,5 @@
+package com.filmdoms.community.article.data.constant;
+
+public enum PostStatus {
+    ACTIVE, INACTIVE
+}
