--- src/main/java/com/filmdoms/community/board/data/constant/PostStatus.java
@@ -1,5 +1,0 @@
-package com.filmdoms.community.board.data.constant;
-
-public enum PostStatus {
-    ACTIVE, INACTIVE
-}
