+++ src/main/java/com/filmdoms/community/comment/data/dto/constant/CommentStatus.java
@@ -1,4 +1,4 @@
-package com.filmdoms.community.board.data.constant;
+package com.filmdoms.community.comment.data.dto.constant;
 
 public enum CommentStatus {
     ACTIVE, INACTIVE, DELETED;
